+++ tests/tests-framework/src/main/java/org/eclipse/dirigible/tests/IntegrationTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2025 Eclipse Dirigible contributors
+ * Copyright (c) 2025 Eclipse Dirigible contributors
  *
  * All rights reserved. This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v2.0 which accompanies this distribution, and is available at
@@ -7,20 +7,47 @@
  *
  * SPDX-FileCopyrightText: Eclipse Dirigible contributors SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.dirigible.integration.tests;
+package org.eclipse.dirigible.tests;
 
 import org.awaitility.Awaitility;
-import org.eclipse.dirigible.tests.DirigibleTestTenant;
+import org.eclipse.dirigible.commons.config.Configuration;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.annotation.DirtiesContext.ClassMode;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-public abstract class DirigibleIntegrationTest extends IntegrationTest {
+@AutoConfigureMockMvc
+@DirtiesContext(classMode = ClassMode.AFTER_EACH_TEST_METHOD)
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+public abstract class IntegrationTest {
+
+    // set config to false if you want to disable the headless mode
+    static {
+        com.codeborne.selenide.Configuration.headless = true;
+    }
 
     @Autowired
+    private DirigibleCleaner dirigibleCleaner;
+    @Autowired
     private TenantCreator tenantCreator;
+
+    @AfterEach
+    final void cleanUp() {
+        dirigibleCleaner.clean();
+    }
+
+    @AfterAll
+    public static final void reloadConfigurations() {
+        Configuration.reloadConfigurations();
+    }
 
     protected void createTenants(DirigibleTestTenant... tenants) {
         createTenants(Arrays.asList(tenants));
@@ -40,4 +67,5 @@ public abstract class DirigibleIntegrationTest extends IntegrationTest {
                   .atMost(35, TimeUnit.SECONDS)
                   .until(() -> tenantCreator.isTenantProvisioned(tenant));
     }
+
 }
