--- tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/IntegrationTest.java
@@ -1,46 +1,0 @@
-/*
- * Copyright (c) 2025 Eclipse Dirigible contributors
- *
- * All rights reserved. This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v2.0 which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v20.html
- *
- * SPDX-FileCopyrightText: Eclipse Dirigible contributors SPDX-License-Identifier: EPL-2.0
- */
-package org.eclipse.dirigible.integration.tests;
-
-import org.eclipse.dirigible.commons.config.Configuration;
-import org.eclipse.dirigible.tests.DirigibleCleaner;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.AfterEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.annotation.DirtiesContext.ClassMode;
-
-@AutoConfigureMockMvc
-@DirtiesContext(classMode = ClassMode.AFTER_EACH_TEST_METHOD)
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-public abstract class IntegrationTest {
-
-    // set config to false if you want to disable the headless mode
-    static {
-        com.codeborne.selenide.Configuration.headless = true;
-    }
-
-    @Autowired
-    private DirigibleCleaner dirigibleCleaner;
-
-    @AfterEach
-    final void cleanUp() {
-        dirigibleCleaner.clean();
-    }
-
-    @AfterAll
-    public static final void reloadConfigurations() {
-        Configuration.reloadConfigurations();
-    }
-
-}
