+++ tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/api/rest/ODataAPIIT.java
@@ -10,7 +10,7 @@
 package org.eclipse.dirigible.integration.tests.api.rest;
 
 import io.restassured.http.ContentType;
-import org.eclipse.dirigible.integration.tests.IntegrationTest;
+import org.eclipse.dirigible.tests.IntegrationTest;
 import org.eclipse.dirigible.tests.restassured.RestAssuredExecutor;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
