+++ tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/ui/tests/BPMStarterTemplateIT.java
@@ -11,8 +11,8 @@ package org.eclipse.dirigible.integration.tests.ui.tests;
 
 import ch.qos.logback.classic.Level;
 import org.eclipse.dirigible.components.api.bpm.BpmFacade;
-import org.eclipse.dirigible.integration.tests.UserInterfaceIntegrationTest;
 import org.eclipse.dirigible.tests.FormView;
+import org.eclipse.dirigible.tests.UserInterfaceIntegrationTest;
 import org.eclipse.dirigible.tests.WelcomeView;
 import org.eclipse.dirigible.tests.Workbench;
 import org.eclipse.dirigible.tests.framework.HtmlElementType;
