+++ tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/ui/tests/DeclineLeaveRequestBpmIT.java
@@ -9,7 +9,7 @@
  */
 package org.eclipse.dirigible.integration.tests.ui.tests;
 
-import org.eclipse.dirigible.integration.tests.PredefinedProjectIT;
+import org.eclipse.dirigible.tests.PredefinedProjectIT;
 import org.eclipse.dirigible.tests.mail.GreenMailConfig;
 import org.eclipse.dirigible.tests.projects.TestProject;
 import org.springframework.beans.factory.annotation.Autowired;
