+++ tests/tests-integrations/src/test/java/org/eclipse/dirigible/integration/tests/ui/tests/MultitenancyIT.java
@@ -10,8 +10,8 @@
 package org.eclipse.dirigible.integration.tests.ui.tests;
 
 import org.eclipse.dirigible.commons.config.DirigibleConfig;
-import org.eclipse.dirigible.integration.tests.UserInterfaceIntegrationTest;
 import org.eclipse.dirigible.tests.DirigibleTestTenant;
+import org.eclipse.dirigible.tests.UserInterfaceIntegrationTest;
 import org.eclipse.dirigible.tests.framework.Browser;
 import org.eclipse.dirigible.tests.framework.BrowserFactory;
 import org.eclipse.dirigible.tests.framework.HtmlElementType;
